Add authentication request headers to RPC call
